requirements.txt
flask==1.1.2
pillow==8.2.0
requests==2.25.1